Improve code organization and terminology
Split lib.rs into two files, one for the client and the server. Also
starts to consistently use the terms "client" and "server" rather then
"acceptor" and "connector".

Removes the "legacy" functions from the api, instead default to the
(not spec-conforming) behaviour of the reference implementation.
A spec-conforming implementation will be released as a new version
when the reference implementation is fixed.